requirements.txt
Django==2.2.3
djangorestframework==3.7.7
get==2019.4.13
post==2019.4.13
public==2019.4.13
pytz==2019.1
query-string==2019.4.13
request==2019.4.13
sqlparse==0.3.0
virtualenv==16.6.1
virtualenvwrapper-win==1.2.5
